modified the layout for home page and login.
Alex was wondering if we really need separate log in for prof and
student. We can have them both log in at the same screen and then
depending on their role direct them to the correct page
Sweta thought we can use two login.php papges and each one will have a
hardcoded value for login user type.. like Professor or ...Student..
then redirect.
Alex suggested that doing it that way is unnecessary because when they
login we can just check their role and re direct from there. there's
really no point in having a separate login.
Sweta agreed and noted to modify the layout for home page and login.